Added a class check for analytics
Added a class check for the analytics bundle in the disciple-tools.php
file. It just checks if the GA plugin is enabled and disables our
bundle so there is no crash conflict. We should build an alert that
requests that they disable the GA plugin for our plugin to work
properly.