Rename XXX-controller.php to XXX.php
Because Git doesn't track renames, but only sees file additions and file
deletions, it may guess wrong as to which file got renamed to what.

Here is my summary:

    renamed:    dt-asset-mapping/asset-mapping-controller.php -> dt-asset-mapping/asset-mapping.php
    renamed:    dt-contacts/contacts-controller.php -> dt-contacts/contacts.php
    renamed:    dt-groups/groups-controller.php -> dt-groups/groups.php
    renamed:    dt-locations/locations-controller.php -> dt-locations/locations.php
    renamed:    dt-people-groups/people-groups-controller.php -> dt-people-groups/people-groups.php
    renamed:    dt-prayer/prayer-controller.php -> dt-prayer/prayer.php
    renamed:    dt-resources/resources-controller.php -> dt-resources/resources.php
    renamed:    dt-statistics/statistics-controller.php -> dt-statistics/statistics.php
    renamed:    dt-users/users-controller.php -> dt-users/users.php

The class names were renamed too, and the references were renamed.

The reason for the rename was that we weren't using the word controller
in strictly the same fashion as in the MVC design pattern, nor in the
way Wordpress documentation refers to controllers. We decided to avoid
MVC terminology.