changed table column from item_id to post_id
I changed the table column from item_it to post_id to be more clear,
and I refactored all relevant files. The reasoning behind this was that
in an @mention you had a relevant comment_id, but in the field update
you had a less relevant meta_id to store, but in all cases you had the
post_id. So it was confusing to have the post_id labeled as secondary
and the meta/comment id labeled as primary. I think this will lead to
less confusion as we extend the section.